backend/requirements.txt
flask
flask-cors
flask-jwt-extended
pyodbc
bcrypt
pandas
numpy
google-auth
requests 
python-dotenv
gunicorn